Restructuring the api with flask blueprints and separating the schemas in their own script
