+++ flaskTodoAPi/__init__.py
@@ -7,6 +7,7 @@ from flask_marshmallow import Marshmallow
 from flask_cors import CORS
 
 
+
 app = Flask(__name__)
 CORS(app, resources='/api/*')
 basedir = os.path.abspath(os.path.dirname(__file__)) 
@@ -18,4 +19,12 @@ migrate = Migrate(app,db)
 
 ma = Marshmallow(app) 
 
-from flaskTodoAPi import routes
+#from flaskTodoAPi import routes
+
+from flaskTodoAPi.users.routes import users
+from flaskTodoAPi.tasks.routes import tasks
+from flaskTodoAPi.tags.routes import tags
+
+app.register_blueprint(users)
+app.register_blueprint(tasks)
+app.register_blueprint(tags)
